Refactoring. Moved Listeners to gdx, moved ProgramCardsView and UI back to the ui package. Refactored removal of conflicting positions, refactored to make fun mode work again.
