+++ src/main/java/roborally/ui/gdx/listeners/LaserListener.java
@@ -1,4 +1,4 @@
-package roborally.ui.listeners;
+package roborally.ui.gdx.listeners;
 
 import com.badlogic.gdx.math.GridPoint2;
 import roborally.game.gameboard.objects.laser.LaserRegister;
