+++ src/main/java/roborally/ui/gdx/listeners/WallListener.java
@@ -1,4 +1,4 @@
-package roborally.ui.listeners;
+package roborally.ui.gdx.listeners;
 
 import com.badlogic.gdx.math.GridPoint2;
 import roborally.ui.ILayers;
